[JBAOP-666] Rework classpools to always delegate to the domain and to use the class resource name to determine if the class can be loaded.
Better handling of parentLast domains.

Remove AOPClassPool.registerGeneratedClass() and the calls to that from TransformerCommon.toClass(). Register these when caching the class from ClassPool.make() if dynamic, and also put them into the pools local resources so the domains can find them.

Exclude the test classes from the -core .classpath

Modify AOPClasPool.isLocalResource() to take a resource path instead of the raw classname and modify the callers.

Modify ClassPoolDomain to use a ClassPoolToClassPoolDomainAdaptor as its parent if the domain's parent is null.



git-svn-id: https://svn.jboss.org/repos/jbossas/projects/aop/trunk@82157 84be2c1e-ba19-0410-b317-a758671a6fc1